Use ADD for job recency indicator.
Marty's suggestion was a pretty good one here.  Just use ADD with
expiration and let the server sort it out.  This prevents there from
being a client clock skew issue causing jobs to run too frequently,
but does mean we don't have a record of the most recent run of a job.
Of course, we could just stick that in another key.