Dead node cleaner.
Closes #11

After nodes stop checking in, we stop advertising availability as a
blob service.

This requires a view called cbfs/nodes that looks like this:

function (doc, meta) {
  if (doc.type === "storage") {
    emit(doc.time, null);
  }
}

and a view called cbs/node_blobs that looks like this:

function (doc, meta) {
  if (doc.type === "blob") {
    for (var n in doc.nodes) {
      emit(n, null);
    }
  }
}

I added a _count reducer to mine, but it's not required.