More conservative GC.
Update timestamps whenever something is used and don't GC anything
that's been used in the last hour.

Both sides cooperate so that even if one begins to slip through, or a
request comes in late, the deleting side will override the request and
not delete something that's been recently referenced.

The deletion can probably still be pulled out of a CAS loop so there's
only one attempt, but this closes a window.

Closes #83