Reserve space immediately upon queueing blob request.
This allows us to start rejecting blobs faster during a storm of
incoming requests while we're already approaching limits.

I'm not entirely sure this solves the observed issue, but it should
definitely slow it down while the view catches up.  The other nodes
won't know that we're as full as we think we are until the transfers
complete *and* the view updates.  When the view runs again locally,
we'll forget what we have in transfer, but it is expected that the
requests will come in bursts and generally things will catch up before
it matters again.

cbugg: close bug-270