mortem in quisquilias
When dealing with relocating blobs from a dying node, if the blob is
considered trash, just let it be collected from the dead node and don't
worry about trying to replicate it.

I removed a node from my cluster and the cluster refused to let go of
the node because it was the only node containing a few thousand blobs
that were considered garbage.  The node was gone, so it was not
recoverable, but the blobs were garbage, so we didn't want them anyway.